+++ uni_modules/flower-compound/components/ui-layout-source/ui-layout-source.vue
@@ -201,14 +201,5 @@
 </script>
 
 <style scoped>
-	._ui-layout-source__empty-content {
-		position: absolute;
-		left: 0;
-		right: 0;
-		top: 0;
-		bottom: 0;
-		display: flex;
-		justify-content: center;
-		align-items: center;
-	}
+	
 </style>