+++ uni_modules/flower-config/core/theme.ts
@@ -6,6 +6,8 @@
 import { reactive } from "vue";
 import { generate } from "../plug-in/design-colors";
 import { defaultPrimary, defaultColors, defaultPresets, defaultInternalSetup } from "./mergeConfiguration";
+import { switchGlobalModel } from "./darkThemeModel";
+import { getCurrentPageRoute } from "./pages"
 
 // 颜色集
 let colors : object = reactive({});
@@ -23,6 +25,22 @@ const defaultColorGenerate = (colorKey : string, model : "light" | "dark" = "lig
 };
 
 /**
+ * 初始化主题深色/浅色模式
+ * @param {String} model ["light","dark"] 模式
+ */
+const initThemeModel = () : void => {
+	// 缓存主题色
+	const model = defaultInternalSetup.model;
+	// #ifdef APP-PLUS
+	setPullToRefreshColor(defaultColors.primary);
+	// #endif
+	// 初始化颜色
+	for (let i in defaultColors) { colors = Object.assign(colors, defaultColorGenerate(i, model)); };
+	// 预设颜色
+	for (let p in defaultPresets) { if (typeof defaultPresets[p] == "string") { colors[p] = defaultPresets[p]; } else { colors[p] = defaultPresets[p][model == "light" ? 0 : 1]; }; };
+};
+
+/**
  * 设置深色/浅色模式
  * @param {String} model ["light","dark"] 模式
  */
@@ -30,11 +48,21 @@ const setThemeModel = (model : "light" | "dark" = defaultInternalSetup.model) : 
 	// 缓存主题色
 	defaultInternalSetup.model = model;
 	uni.setStorageSync("flower-library-theme-model", model);
-	setPullToRefreshColor(defaultColors.primary);
-	// 初始化颜色
-	for (let i in defaultColors) { colors = Object.assign(colors, defaultColorGenerate(i, model)); };
-	// 预设颜色
-	for (let p in defaultPresets) { if (typeof defaultPresets[p] == "string") { colors[p] = defaultPresets[p]; } else { colors[p] = defaultPresets[p][model == "light" ? 0 : 1]; }; };
+	// #ifdef APP-PLUS
+	switchGlobalModel(model == "dark" ? true : false, getCurrentPageRoute());
+	// #endif
+	// 重置初始化
+	initThemeModel();
+};
+
+/**
+ * 获取暗黑模式
+ */
+const getThemeModel = () => {
+	return defaultInternalSetup.model;
+};
+const getThemeModelBoolean = () => {
+	return defaultInternalSetup.model == "dark" ? true : false;
 };
 
 /**
@@ -56,22 +84,36 @@ const restoreThemeColor = () => {
 	setPullToRefreshColor("defaultPrimary");
 	colors = Object.assign(colors, defaultColorGenerate("primary", defaultInternalSetup.model));
 };
-
+/**
+ * 获取主题
+ */
+const getThemeColor = () => {
+	return defaultColors.primary;
+};
+/**
+ * 设置默哀模式
+ */
+const setSacrificeMode = (mode : boolean = false) => {
+	defaultInternalSetup.sacrifice = mode;
+	uni.setStorageSync("flower-library-sacrifice-model", mode);
+};
+/**
+ * 获取默哀模式值
+ */
+const getSacrificeMode = () => {
+	return defaultInternalSetup.sacrifice;
+}
 /**
  * 当未配置下拉刷新颜色时，自动装配主题色
  * @param {String} color 主题颜色
  */
 const setPullToRefreshColor = (color : string) => {
-	// #ifdef APP-PLUS
 	if (!__uniConfig.globalStyle.pullToRefresh) {
 		__uniConfig.globalStyle.pullToRefresh = {
 			color: color
 		};
 	};
-	// #endif
 };
-
-setPullToRefreshColor("#1890ff");
 
 /**
  * 参数是否为内置变量，如果不是则直接返回
@@ -82,10 +124,14 @@ const isBuiltColor = (color : string) : string => {
 };
 /**
  * 获取 flower-library 颜色
- * @param {String} color  颜色值 / 内置颜色
+ * @param {String,Array} color  颜色值 / 内置颜色
  */
-const getColors = (color : string) : string => {
-	return isBuiltColor(color)
+const getColors = (color : string | any[]) : string => {
+	if (typeof color == "object") {
+		return getThemeModelBoolean() ? color[1] : color[0]
+	} else {
+		return isBuiltColor(color)
+	};
 };
 /**
  * 获取 flower-library 颜色
@@ -111,6 +157,13 @@ const setFontSizeIncrease = (parameter : number) : void => {
 	uni.setStorageSync("flower-library-fontszie-increase", parameter)
 };
 /**
+ * 获取字体大小值
+ */
+const getFontSizeIncrease = () => {
+	return defaultInternalSetup.increase;
+}
+
+/**
  * 重置字体大小设置
  */
 const restoreFontSizeIncrease = () : void => {
@@ -121,13 +174,20 @@ const restoreFontSizeIncrease = () : void => {
 export {
 	generate,
 	defaultColors,
+	initThemeModel,
 	setThemeModel,
+	getThemeModel,
+	getThemeModelBoolean,
 	setThemeColor,
+	getThemeColor,
 	restoreThemeColor,
 	setFontSizeIncrease,
+	getFontSizeIncrease,
 	restoreFontSizeIncrease,
 	isBuiltColor,
 	getColors,
 	getTextColors,
-	getBackgroundColors
+	getBackgroundColors,
+	setSacrificeMode,
+	getSacrificeMode
 }